requirements.txt
crispy-bootstrap5
Django
django-crispy-forms
django-import-export
django-registration-redux
django-tables2
django-tinymce
python-decouple
django-filter
django-versatileimagefield